Use static manifest
